requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Google AI
google-genai>=0.1.0

# Telegram
python-telegram-bot>=21.0

# Supabase / PostgreSQL
supabase>=2.0.0

# Google Calendar
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# Notion
notion-client>=2.0.0

# Redis (session)
redis>=5.0.0

# HTTP client
httpx>=0.27.0

# Audio
openai-whisper>=20231117  # alternativa a Gemini para transcripción

# Dev / Test (no van en producción)
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
